+++ loader.py
@@ -4,7 +4,7 @@ import random
 import torch
 import time
 from tqdm import tqdm
-from definitions import *
+#from definitions import *
 from torch.utils.data import Dataset, DataLoader
 
 INPUT_METHOD_ALL_CONCAT = "concat_defs"
@@ -34,10 +34,10 @@ class DefinitionsDataset(Dataset):
         split_line = line.split()
         word = split_line[0]
         if self.input_method == INPUT_METHOD_ONE:
-		definition = get_a_definition(word)
+            definition = get_a_definition(word)
         elif self.input_method == INPUT_METHOD_ALL_CONCAT:
-		definition = get_definitions_concat(word)
-        embedding = np.array([float(val) for val in split_line[1:]])
+            definition = get_definitions_concat(word)
+            embedding = np.array([float(val) for val in split_line[1:]])
         return word, definition, embedding
 
     def __getitem__(self, idx):
