+++ model/__init__.py
@@ -1,4 +1,8 @@
+from .Seq2SeqModel import Seq2SeqModel
+from .Def2VecModel import Def2VecModel
+from .BaselineModel import BaselineModel
 from .EncoderRNN import EncoderRNN
 from .DecoderRNN import DecoderRNN
+
 #from .TopKDecoder import TopKDecoder
 #from .seq2seq import Seq2seq
