+++ scripts/nmt_glove_combine.py
@@ -1,9 +1,5 @@
-from __future__ import print_function
-import collections
-import torch
-import torch.optim as optim
-import torch.nn as nn
 import numpy as np
+import torch.nn as nn
 import torchtext.vocab as vocab
 import argparse
 import shutil
