+++ scripts/nmt_glove_visualize.py
@@ -1,18 +1,13 @@
-from __future__ import print_function
-import collections
-import torch
-import torch.optim as optim
-import torch.nn as nn
 import numpy as np
+import torch
+import torch.nn as nn
 import torchtext.vocab as vocab
 import argparse
-import shutil
 from tqdm import tqdm
 from model import Def2VecModel
 from torch.autograd import Variable
 from loader import *
-from tensorboardX import SummaryWriter
-from pytorch_monitor import monitor_module, init_experiment
+from pytorch_monitor import init_experiment
 
 CONFIG = dict(
         title="def2vec",
